[fix] google engine: Result image thumbnails
